jimt/CT-446 - fileUploadErrorCallback
