+++ apps/src/codebridge/FileBrowser/hooks/index.ts
@@ -1,1 +1,2 @@
+export * from './useFileUploadErrorCallback';
 export * from './useHandleFileUpload';
